+++ lib/ui/screens/login_screen/widgets/account_creating_content/index.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:mooncake/ui/ui.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 
 class CreatingAccountPopupContent extends StatelessWidget {
   @override
@@ -11,17 +12,37 @@ class CreatingAccountPopupContent extends StatelessWidget {
           PostsLocalizations.of(context)
               .translate(Messages.creatingAccountPopupTitle)
               .toUpperCase(),
-          style: Theme.of(context).accentTextTheme.headline6,
+          style: Theme.of(context)
+              .textTheme
+              .headline6
+              .copyWith(color: Theme.of(context).colorScheme.primary),
         ),
         SizedBox(height: 10),
         Text(
           PostsLocalizations.of(context)
               .translate(Messages.creatingAccountText),
-          style: Theme.of(context).accentTextTheme.headline6,
+          style: TextStyle(
+            color: Theme.of(context).colorScheme.primary,
+          ),
         ),
         SizedBox(height: 30),
-        LoadingIndicator(),
+        LoadingIndicator(color: Theme.of(context).colorScheme.secondaryVariant),
+        SizedBox(height: 30),
+        SizedBox(
+          width: double.infinity,
+          child: SecondaryLightInvertRoundedButton(
+            child: Text(
+              PostsLocalizations.of(context).translate(Messages.cancel),
+              textAlign: TextAlign.center,
+            ),
+            onPressed: () => _handleCancel(context),
+          ),
+        ),
       ],
     );
+  }
+
+  void _handleCancel(BuildContext context) {
+    BlocProvider.of<AccountBloc>(context).add(LogOut());
   }
 }